+++ packages/components/src/utils/types.ts
@@ -57,3 +57,11 @@ export type ComponentWithAs<Component extends As, Props extends object = {}> = {
   defaultProps?: Partial<any>;
   id?: string;
 };
+
+export type ForwardRefWithStaticComponents<
+  Props extends Record<string, any>,
+  Static extends Record<string, any>
+> = ((props: Props) => React.ReactElement) &
+  Static & {
+    displayName: string;
+  };