+++ packages/components/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @kubed/components
 
+## 0.0.41
+
+### Patch Changes
+
+- [`e76b9e6`](https://github.com/kubesphere/kube-design/commit/e76b9e61580072d858c46645bfad821219900c2f) Thanks [@donniean](https://github.com/donniean)! - feature improvements and bug fixes
+
+## 0.0.40
+
+### Patch Changes
+
+- [`f13f2b4`](https://github.com/kubesphere/kube-design/commit/f13f2b4b3d26547e1e11e0438db8e0bba8a60662) Thanks [@chenz24](https://github.com/chenz24)! - Refine export.
+
+## 0.0.39
+
+### Patch Changes
+
+- [`31ff604`](https://github.com/kubesphere/kube-design/commit/31ff6042e5b1f333941c7f5ef336ced794968143) Thanks [@chenz24](https://github.com/chenz24)! - 1 fix bugs
+
+## 0.0.38
+
+### Patch Changes
+
+- [`2246672`](https://github.com/kubesphere/kube-design/commit/2246672041f06277c9c90d2e9187817a460baeca) Thanks [@chenz24](https://github.com/chenz24)! - 1. fix: Add missing z-index to TypeSelect.
+  2. fix: Export missing component of ShowMore.
+  3. feat: Refine style of tab component
+
+* [`4620226`](https://github.com/kubesphere/kube-design/commit/462022690696387ffce187b6981f5dfe5cbf52e3) Thanks [@chenz24](https://github.com/chenz24)! - 1. feat: Add DownCircle,Forward,PlugCircle icons
+  2. fix: Refine disabled style of button.
+* Updated dependencies [[`2246672`](https://github.com/kubesphere/kube-design/commit/2246672041f06277c9c90d2e9187817a460baeca)]:
+  - @kubed/icons@0.0.3
+
+## 0.0.37
+
+### Patch Changes
+
+- [`93a25cd`](https://github.com/kubesphere/kube-design/commit/93a25cd44eef2ec3076421d50f8f08a2a96e9b2c) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add ShowMore component.
+  2. Tag support closable.
+
+## 0.0.36
+
+### Patch Changes
+
+- [`c15bc73`](https://github.com/kubesphere/kube-design/commit/c15bc73b6ea45e55e74cf4a7c084c7f3d70305f0) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add TypeSelect component.
+  2. Add Steps component.
+  3. Add another style of loading.
+  4. Fix some bugs.
+
 ## 0.0.35
 
 ### Patch Changes